[Search] Move `ConnectorDefinition` types and consts (#197676)
## Summary

- Move types and consts for `ConnectorClientSideDefinition` and
`ConnectorServerSideDefinition` to the shared `kbn-search-connectors`
package
- Update ESS references to these values to use it from the package
- Remove them from the connectors plugin

(cherry picked from commit 8ed8cc964e2ec557ea7173cc8cecf17b7f69d0d3)
