+++ packages/kbn-search-connectors/index.ts
@@ -16,6 +16,7 @@ export const CONNECTORS_ACCESS_CONTROL_INDEX_PREFIX = '.search-acl-filter-';
 export const CRAWLER_SERVICE_TYPE = 'elastic-crawler';
 
 export * from './components';
+export * from './constants';
 export * from './lib';
 export * from './types';
 export * from './utils';
