+++ packages/kbn-search-connectors/types/native_connectors.ts
@@ -11,54 +11,64 @@ import { i18n } from '@kbn/i18n';
 
 import { DisplayType, FeatureName, FieldType, NativeConnector } from './connectors';
 
-const USERNAME_LABEL = i18n.translate('searchConnectors.nativeConnectors.usernameLabel', {
+// assigning these to a local var significantly improves bundle size
+// because it reduces references to the imported modules.
+const { translate } = i18n;
+const { TEXTBOX, TEXTAREA, NUMERIC, TOGGLE, DROPDOWN } = DisplayType;
+const {
+  SYNC_RULES,
+  INCREMENTAL_SYNC,
+  DOCUMENT_LEVEL_SECURITY,
+  FILTERING_ADVANCED_CONFIG,
+  FILTERING_RULES,
+} = FeatureName;
+const { STRING, LIST, INTEGER, BOOLEAN } = FieldType;
+
+const USERNAME_LABEL = translate('searchConnectors.nativeConnectors.usernameLabel', {
   defaultMessage: 'Username',
 });
 
-const PASSWORD_LABEL = i18n.translate('searchConnectors.nativeConnectors.passwordLabel', {
+const PASSWORD_LABEL = translate('searchConnectors.nativeConnectors.passwordLabel', {
   defaultMessage: 'Password',
 });
 
-const ENABLE_SSL_LABEL = i18n.translate('searchConnectors.nativeConnectors.enableSSL.label', {
+const ENABLE_SSL_LABEL = translate('searchConnectors.nativeConnectors.enableSSL.label', {
   defaultMessage: 'Enable SSL',
 });
 
-const SSL_CERTIFICATE_LABEL = i18n.translate(
-  'searchConnectors.nativeConnectors.sslCertificate.label',
-  {
-    defaultMessage: 'SSL certificate',
-  }
-);
+const SSL_CERTIFICATE_LABEL = translate('searchConnectors.nativeConnectors.sslCertificate.label', {
+  defaultMessage: 'SSL certificate',
+});
 
-const RETRIES_PER_REQUEST_LABEL = i18n.translate(
+const RETRIES_PER_REQUEST_LABEL = translate(
   'searchConnectors.nativeConnectors.retriesPerRequest.label',
   {
     defaultMessage: 'Retries per request',
   }
 );
 
-const ADVANCED_RULES_IGNORED_LABEL = i18n.translate(
+const ADVANCED_RULES_IGNORED_LABEL = translate(
   'searchConnectors.nativeConnectors.advancedRulesIgnored.label',
   {
     defaultMessage: 'This configurable field is ignored when Advanced Sync Rules are used.',
   }
 );
 
-const MAX_CONCURRENT_DOWNLOADS_LABEL = i18n.translate(
+const MAX_CONCURRENT_DOWNLOADS_LABEL = translate(
   'searchConnectors.nativeConnectors.nativeConnectors.maximumConcurrentLabel',
   {
     defaultMessage: 'Maximum concurrent downloads',
   }
 );
 
-const USE_TEXT_EXTRACTION_SERVICE_LABEL = i18n.translate(
+const USE_TEXT_EXTRACTION_SERVICE_LABEL = translate(
   'searchConnectors.nativeConnectors.textExtractionService.label',
   {
     defaultMessage: 'Use text extraction service',
   }
 );
 
-const USE_TEXT_EXTRACTION_SERVICE_TOOLTIP = i18n.translate(
+const USE_TEXT_EXTRACTION_SERVICE_TOOLTIP = translate(
   'searchConnectors.nativeConnectors.textExtractionService.tooltip',
   {
     defaultMessage:
@@ -67,7 +77,7 @@ const USE_TEXT_EXTRACTION_SERVICE_TOOLTIP = i18n.translate(
   }
 );
 
-const ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL = i18n.translate(
+const ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL = translate(
   'searchConnectors.nativeConnectors.enableDLS.label',
   {
     defaultMessage: 'Enable document level security',
@@ -75,21 +85,21 @@ const ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL = i18n.translate(
 );
 
 const getEnableDocumentLevelSecurityTooltip = (serviceName: string) =>
-  i18n.translate('searchConnectors.nativeConnectors.enableDLS.tooltip', {
+  translate('searchConnectors.nativeConnectors.enableDLS.tooltip', {
     defaultMessage:
       'Document level security ensures identities and permissions set in {serviceName} are maintained in Elasticsearch. This enables you to restrict and personalize read-access users and groups have to documents in this index. Access control syncs ensure this metadata is kept up to date in your Elasticsearch documents.',
     values: { serviceName },
   });
 
-const DATABASE_LABEL = i18n.translate('searchConnectors.nativeConnectors.databaseLabel', {
+const DATABASE_LABEL = translate('searchConnectors.nativeConnectors.databaseLabel', {
   defaultMessage: 'Database',
 });
 
-const SCHEMA_LABEL = i18n.translate('searchConnectors.nativeConnectors.schemaLabel', {
+const SCHEMA_LABEL = translate('searchConnectors.nativeConnectors.schemaLabel', {
   defaultMessage: 'Schema',
 });
 
-const PORT_LABEL = i18n.translate('searchConnectors.nativeConnectors.portLabel', {
+const PORT_LABEL = translate('searchConnectors.nativeConnectors.portLabel', {
   defaultMessage: 'Port',
 });
 
@@ -103,19 +113,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       account_name: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.azureBlobStorage.accountNameLabel',
-          {
-            defaultMessage: 'Account name',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.azureBlobStorage.accountNameLabel', {
+          defaultMessage: 'Account name',
+        }),
         options: [],
         order: 1,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -123,19 +130,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       account_key: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.azureBlobStorage.accountKeyLabel',
-          {
-            defaultMessage: 'Account key',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.azureBlobStorage.accountKeyLabel', {
+          defaultMessage: 'Account key',
+        }),
         options: [],
         order: 2,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -143,20 +147,17 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       blob_endpoint: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.azureBlobStorage.blobEndpointLabel',
-          {
-            defaultMessage: 'Blob endpoint',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.azureBlobStorage.blobEndpointLabel', {
+          defaultMessage: 'Blob endpoint',
+        }),
         options: [],
         order: 3,
         placeholder: 'http://127.0.0.1:10000/devstoreaccount',
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -164,19 +165,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       containers: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.azureBlobStorage.containerNameLabel',
-          {
-            defaultMessage: 'List of containers',
-          }
-        ),
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.azureBlobStorage.containerNameLabel', {
+          defaultMessage: 'List of containers',
+        }),
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -184,14 +182,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 5,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -199,14 +197,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 100,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: MAX_CONCURRENT_DOWNLOADS_LABEL,
         options: [],
         order: 6,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [
           {
@@ -219,29 +217,29 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: false },
         basic: { enabled: true },
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.azureBlobStorage.name', {
+    name: translate('searchConnectors.nativeConnectors.azureBlobStorage.name', {
       defaultMessage: 'Azure Blob Storage',
     }),
     serviceType: 'azure_blob_storage',
@@ -251,18 +249,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       path: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.box.pathLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.box.pathLabel', {
           defaultMessage: 'Path to fetch files/folders',
         }),
         options: [],
         order: 1,
         required: false,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.box.pathTooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.box.pathTooltip', {
           defaultMessage: 'Path is ignored when Advanced Sync Rules are used. ',
         }),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -270,8 +268,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       app_key: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.box.appKeyLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.box.appKeyLabel', {
           defaultMessage: 'App Key',
         }),
         options: [],
@@ -279,7 +277,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -287,8 +285,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       app_secret: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.box.appSecretLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.box.appSecretLabel', {
           defaultMessage: 'App secret',
         }),
         options: [],
@@ -296,7 +294,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -304,8 +302,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       refresh_token: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.box.refreshTokenLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.box.refreshTokenLabel', {
           defaultMessage: 'Refresh token',
         }),
         options: [],
@@ -313,7 +311,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -321,14 +319,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 5,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -336,14 +334,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 100,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: MAX_CONCURRENT_DOWNLOADS_LABEL,
         options: [],
         order: 6,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -351,14 +349,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
@@ -366,18 +364,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 8,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.boxTooltip.name', {
+          translate('searchConnectors.nativeConnectors.boxTooltip.name', {
             defaultMessage: 'Box',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -390,29 +388,26 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate('searchConnectors.nativeConnectors.box.includeInheritedUsersLabel', {
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.box.includeInheritedUsersLabel', {
           defaultMessage: 'Include groups and inherited users',
         }),
         options: [],
         order: 9,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
-          'searchConnectors.nativeConnectors.box.includeInheritedUsersTooltip',
-          {
-            defaultMessage:
-              'Include groups and inherited users when indexing permissions. Enabling this configurable field will cause a significant performance degradation.',
-          }
-        ),
-        type: FieldType.BOOLEAN,
+        tooltip: translate('searchConnectors.nativeConnectors.box.includeInheritedUsersTooltip', {
+          defaultMessage:
+            'Include groups and inherited users when indexing permissions. Enabling this configurable field will cause a significant performance degradation.',
+        }),
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
       },
     },
     features: {},
-    name: i18n.translate('searchConnectors.nativeConnectors.box.name', {
+    name: translate('searchConnectors.nativeConnectors.box.name', {
       defaultMessage: 'Box',
     }),
     serviceType: 'box',
@@ -422,25 +417,25 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       data_source: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate('searchConnectors.nativeConnectors.confluenceSource.label', {
+        display: DROPDOWN,
+        label: translate('searchConnectors.nativeConnectors.confluenceSource.label', {
           defaultMessage: 'Confluence data source',
         }),
         options: [
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.confluenceCloud.name', {
+            label: translate('searchConnectors.nativeConnectors.confluenceCloud.name', {
               defaultMessage: 'Confluence Cloud',
             }),
             value: 'confluence_cloud',
           },
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.confluenceServer.name', {
+            label: translate('searchConnectors.nativeConnectors.confluenceServer.name', {
               defaultMessage: 'Confluence Server',
             }),
             value: 'confluence_server',
           },
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.confluenceDataCenter.name', {
+            label: translate('searchConnectors.nativeConnectors.confluenceDataCenter.name', {
               defaultMessage: 'Confluence Data Center',
             }),
             value: 'confluence_data_center',
@@ -450,7 +445,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'confluence_server',
@@ -463,8 +458,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'confluence_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.confluenceServer.usernameLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.confluenceServer.usernameLabel', {
           defaultMessage: 'Confluence Server username',
         }),
         options: [],
@@ -472,7 +467,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'admin',
@@ -485,8 +480,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'confluence_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.confluenceServer.passwordLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.confluenceServer.passwordLabel', {
           defaultMessage: 'Confluence Server password',
         }),
         options: [],
@@ -494,7 +489,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -507,19 +502,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'confluence_data_center',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.confluenceDataCenter.usernameLabel',
-          {
-            defaultMessage: 'Username',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.confluenceDataCenter.usernameLabel', {
+          defaultMessage: 'Username',
+        }),
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'admin',
@@ -532,19 +524,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'confluence_data_center',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.confluenceDataCenter.passwordLabel',
-          {
-            defaultMessage: 'Password',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.confluenceDataCenter.passwordLabel', {
+          defaultMessage: 'Password',
+        }),
         options: [],
         order: 5,
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -557,20 +546,17 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'confluence_cloud',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.confluenceCloud.accountEmailLabel',
-          {
-            defaultMessage: 'Confluence Cloud account email',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.confluenceCloud.accountEmailLabel', {
+          defaultMessage: 'Confluence Cloud account email',
+        }),
         options: [],
         order: 6,
         placeholder: 'me@example.com',
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -583,8 +569,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'confluence_cloud',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.confluenceServer.apiTokenLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.confluenceServer.apiTokenLabel', {
           defaultMessage: 'Confluence Cloud API token',
         }),
         options: [],
@@ -592,7 +578,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -600,8 +586,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       confluence_url: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.confluence.urlLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.confluence.urlLabel', {
           defaultMessage: 'Confluence URL label',
         }),
         options: [],
@@ -610,7 +596,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -618,8 +604,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       spaces: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.confluence.spaceKeysLabel', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.confluence.spaceKeysLabel', {
           defaultMessage: 'Confluence space keys',
         }),
         options: [],
@@ -627,7 +613,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: ADVANCED_RULES_IGNORED_LABEL,
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -635,19 +621,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       index_labels: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate('searchConnectors.nativeConnectors.confluence.indexLabelsLabel', {
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.confluence.indexLabelsLabel', {
           defaultMessage: 'Enable indexing labels',
         }),
         options: [],
         order: 10,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.confluence.indexLabelsTooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.confluence.indexLabelsTooltip', {
           defaultMessage:
             'Enabling this will increase the amount of network calls to the source, and may decrease performance',
         }),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -655,14 +641,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_SSL_LABEL,
         options: [],
         order: 11,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -675,14 +661,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SSL_CERTIFICATE_LABEL,
         options: [],
         order: 12,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -690,14 +676,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 13,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -705,14 +691,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 50,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: MAX_CONCURRENT_DOWNLOADS_LABEL,
         options: [],
         order: 14,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [
           {
@@ -730,18 +716,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'confluence_cloud',
           },
         ],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 15,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.confluence.tooltipName', {
+          translate('searchConnectors.nativeConnectors.confluence.tooltipName', {
             defaultMessage: 'Confluence',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -749,32 +735,32 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 16,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.confluence.name', {
+    name: translate('searchConnectors.nativeConnectors.confluence.name', {
       defaultMessage: 'Confluence',
     }),
     serviceType: 'confluence',
@@ -784,14 +770,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       path: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: 'Path to fetch files/folders',
         options: [],
         order: 1,
         required: false,
         sensitive: false,
         tooltip: 'This configurable field is ignored when Advanced Sync Rules are used.',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '/',
@@ -799,7 +785,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       app_key: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: 'App key',
         options: [],
         order: 2,
@@ -807,7 +793,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -815,14 +801,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       app_secret: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: 'App secret',
         options: [],
         order: 3,
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -830,14 +816,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       refresh_token: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: 'Refresh token',
         options: [],
         order: 4,
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -845,14 +831,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: 'Retries per request',
         options: [],
         order: 5,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -860,14 +846,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 100,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: 'Maximum concurrent downloads',
         options: [],
         order: 6,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 100,
@@ -875,14 +861,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
@@ -890,18 +876,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 8,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.dropbox.tooltipName', {
+          translate('searchConnectors.nativeConnectors.dropbox.tooltipName', {
             defaultMessage: 'Dropbox',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -909,8 +895,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       include_inherited_users_and_groups: {
         default_value: null,
         depends_on: [{ field: 'use_document_level_security', value: true }],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.dropbox.includeInheritedUsersAndGroups.label',
           {
             defaultMessage: 'Include groups and inherited users',
@@ -920,32 +906,32 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 9,
         required: false,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.dropbox.includeInheritedUsersAndGroups.tooltip',
           {
             defaultMessage:
               'Include groups and inherited users when indexing permissions. Enabling this configurable field will cause a significant performance degradation.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.dropbox.name', {
+    name: translate('searchConnectors.nativeConnectors.dropbox.name', {
       defaultMessage: 'Dropbox',
     }),
     serviceType: 'dropbox',
@@ -955,19 +941,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       data_source: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.label', {
+        display: DROPDOWN,
+        label: translate('searchConnectors.nativeConnectors.github.label', {
           defaultMessage: 'Data source',
         }),
         options: [
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.github.options.cloudLabel', {
+            label: translate('searchConnectors.nativeConnectors.github.options.cloudLabel', {
               defaultMessage: 'GitHub Cloud',
             }),
             value: 'github_cloud',
           },
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.github.options.cloudServer', {
+            label: translate('searchConnectors.nativeConnectors.github.options.cloudServer', {
               defaultMessage: 'GitHub Server',
             }),
             value: 'github_server',
@@ -977,7 +963,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'github_server',
@@ -990,8 +976,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'github_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.url.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.github.url.label', {
           defaultMessage: 'Server URL',
         }),
         options: [],
@@ -999,7 +985,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1007,13 +993,13 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       auth_method: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.authMethod.label', {
+        display: DROPDOWN,
+        label: translate('searchConnectors.nativeConnectors.github.authMethod.label', {
           defaultMessage: 'Authentication method',
         }),
         options: [
           {
-            label: i18n.translate(
+            label: translate(
               'searchConnectors.nativeConnectors.github.options.personalAccessToken',
               {
                 defaultMessage: 'Personal access token',
@@ -1022,7 +1008,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: PERSONAL_ACCESS_TOKEN,
           },
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.github.options.githubApp', {
+            label: translate('searchConnectors.nativeConnectors.github.options.githubApp', {
               defaultMessage: 'GitHub App',
             }),
             value: GITHUB_APP,
@@ -1032,7 +1018,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: PERSONAL_ACCESS_TOKEN,
@@ -1045,8 +1031,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: PERSONAL_ACCESS_TOKEN,
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.token.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.github.token.label', {
           defaultMessage: 'Token',
         }),
         options: [],
@@ -1054,7 +1040,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1062,19 +1048,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       repo_type: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.repo_type.label', {
+        display: DROPDOWN,
+        label: translate('searchConnectors.nativeConnectors.github.repo_type.label', {
           defaultMessage: 'Repository Type',
         }),
         options: [
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.github.options.organization', {
+            label: translate('searchConnectors.nativeConnectors.github.options.organization', {
               defaultMessage: 'Organization',
             }),
             value: 'organization',
           },
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.github.options.other', {
+            label: translate('searchConnectors.nativeConnectors.github.options.other', {
               defaultMessage: 'Other',
             }),
             value: 'other',
@@ -1083,11 +1069,11 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 5,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.github.repo_type', {
+        tooltip: translate('searchConnectors.nativeConnectors.github.repo_type', {
           defaultMessage:
             'The Document Level Security feature is not available for the Other Repository Type',
         }),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'other',
@@ -1104,8 +1090,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'organization',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.org_name.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.github.org_name.label', {
           defaultMessage: 'Organization Name',
         }),
         options: [],
@@ -1113,7 +1099,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1126,8 +1112,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: GITHUB_APP,
           },
         ],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.appID.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.github.appID.label', {
           defaultMessage: 'App ID',
         }),
         options: [],
@@ -1135,7 +1121,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1148,8 +1134,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: GITHUB_APP,
           },
         ],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.privateKey.label', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.github.privateKey.label', {
           defaultMessage: 'App private key',
         }),
         options: [],
@@ -1157,7 +1143,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1165,18 +1151,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       repositories: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.github.listOfRepos.label', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.github.listOfRepos.label', {
           defaultMessage: 'List of repositories',
         }),
         options: [],
         order: 9,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.github.listOfRepos.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.github.listOfRepos.tooltip', {
           defaultMessage: 'This configurable field is ignored when Advanced Sync Rules are used.',
         }),
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1184,14 +1170,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_SSL_LABEL,
         options: [],
         order: 10,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -1204,14 +1190,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SSL_CERTIFICATE_LABEL,
         options: [],
         order: 11,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1219,14 +1205,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 12,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -1234,14 +1220,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 13,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
@@ -1254,34 +1240,34 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'organization',
           },
         ],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 14,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.github.configuration.useDocumentLevelSecurityTooltip',
           {
             defaultMessage:
               'Document level security ensures identities and permissions set in GitHub are maintained in Elasticsearch. This enables you to restrict and personalize read-access users and groups have to documents in this index. Access control syncs ensure this metadata is kept up to date in your Elasticsearch documents.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.github.name', {
+    name: translate('searchConnectors.nativeConnectors.github.name', {
       defaultMessage: 'Github',
     }),
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
@@ -1292,8 +1278,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       service_account_credentials: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate(
+        display: TEXTAREA,
+        label: translate(
           'searchConnectors.nativeConnectors.gmail.service_account_credentials.label',
           {
             defaultMessage: 'GMail service account JSON',
@@ -1304,7 +1290,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1312,18 +1298,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       subject: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.gmail.subject.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.gmail.subject.label', {
           defaultMessage: 'Google Workspace admin email',
         }),
         options: [],
         order: 2,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.gmail.subject.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.gmail.subject.tooltip', {
           defaultMessage: 'Admin account email address',
         }),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [
           {
@@ -1336,18 +1322,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       customer_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.gmail.customer_id.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.gmail.customer_id.label', {
           defaultMessage: 'Google customer id',
         }),
         options: [],
         order: 3,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.gmail.customer_id.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.gmail.customer_id.tooltip', {
           defaultMessage: 'Google admin console -> Account -> Settings -> Customer Id',
         }),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1355,24 +1341,21 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       include_spam_and_trash: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.gmail.include_spam_and_trash.label',
-          {
-            defaultMessage: 'Include spam and trash emails',
-          }
-        ),
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.gmail.include_spam_and_trash.label', {
+          defaultMessage: 'Include spam and trash emails',
+        }),
         options: [],
         order: 4,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.gmail.include_spam_and_trash.tooltip',
           {
             defaultMessage: 'Will include spam and trash emails, when set to true.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -1380,8 +1363,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.gmail.use_document_level_security.label',
           {
             defaultMessage: 'Enable document level security',
@@ -1391,39 +1374,39 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 5,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.gmail.use_document_level_security.tooltip',
           {
             defaultMessage:
               'Document level security ensures identities and permissions set in GMail are maintained in Elasticsearch. This enables you to restrict and personalize read-access users have to documents in this index. Access control syncs ensure this metadata is kept up to date in your Elasticsearch documents.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.gmail.name', {
+    name: translate('searchConnectors.nativeConnectors.gmail.name', {
       defaultMessage: 'Gmail',
     }),
     serviceType: 'gmail',
   },
   google_cloud_storage: {
     features: {
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
@@ -1431,19 +1414,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       buckets: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.googleCloudStorage.bucketNameLabel',
-          {
-            defaultMessage: 'List of buckets',
-          }
-        ),
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.googleCloudStorage.bucketNameLabel', {
+          defaultMessage: 'List of buckets',
+        }),
         options: [],
         order: 1,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1451,8 +1431,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       service_account_credentials: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.googleCloudStorage.label', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.googleCloudStorage.label', {
           defaultMessage: 'Google Cloud service account JSON',
         }),
         options: [],
@@ -1460,7 +1440,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1468,8 +1448,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.nativeConnectors.googleCloudStorage.retry.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.googleCloudStorage.retry.label', {
           defaultMessage: 'Maximum retries for failed requests',
         }),
         options: [],
@@ -1477,7 +1457,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -1485,20 +1465,20 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
-    name: i18n.translate('searchConnectors.content.nativeConnectors.googleCloud.name', {
+    name: translate('searchConnectors.content.nativeConnectors.googleCloud.name', {
       defaultMessage: 'Google Cloud Storage',
     }),
     serviceType: 'google_cloud_storage',
@@ -1508,19 +1488,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       service_account_credentials: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.gdrive.label', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.gdrive.label', {
           defaultMessage: 'Google Drive service account JSON',
         }),
         options: [],
         order: 1,
         required: true,
         sensitive: true,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.gdrive.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.gdrive.tooltip', {
           defaultMessage:
             'This connectors authenticates as a service account to synchronize content from Google Drive.',
         }),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1528,25 +1508,22 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_domain_wide_delegation_for_sync: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.gdrive.useDomainWideDelegation.label',
-          {
-            defaultMessage: 'Use domain-wide delegation for data sync',
-          }
-        ),
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.gdrive.useDomainWideDelegation.label', {
+          defaultMessage: 'Use domain-wide delegation for data sync',
+        }),
         options: [],
         order: 2,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.gdrive.useDomainWideDelegation.tooltip',
           {
             defaultMessage:
               'Enable domain-wide delegation to automatically sync content from all shared and personal drives in the Google workspace. This eliminates the need to manually share Google Drive data with your service account, though it may increase sync time. If disabled, only items and folders manually shared with the service account will be synced. Please refer to the connector documentation to ensure domain-wide delegation is correctly configured and has the appropriate scopes.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -1559,8 +1536,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.gdrive.workspaceAdminEmailDataSync.label',
           {
             defaultMessage: 'Google Workspace admin email',
@@ -1570,14 +1547,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 3,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.gdrive.workspaceAdminEmailDataSync.tooltip',
           {
             defaultMessage:
               'Provide the admin email to be used with domain-wide delegation for data sync. This email enables the connector to utilize the Admin Directory API for listing organization users. Please refer to the connector documentation to ensure domain-wide delegation is correctly configured and has the appropriate scopes.',
           }
         ),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [
           {
@@ -1595,8 +1572,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.gdrive.workspaceEmailSharedDrivesSync.label',
           {
             defaultMessage: 'Google Workspace email for syncing shared drives',
@@ -1606,14 +1583,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 4,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.gdrive.workspaceEmailSharedDrivesSync.tooltip',
           {
             defaultMessage:
               'Provide the Google Workspace user email for discovery and syncing of shared drives. Only the shared drives this user has access to will be synced.',
           }
         ),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [
           {
@@ -1626,18 +1603,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 5,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.gdrive.tooltipName', {
+          translate('searchConnectors.nativeConnectors.gdrive.tooltipName', {
             defaultMessage: 'Google Drive',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -1654,23 +1631,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: false,
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.gdrive.workspaceAdminEmail.label',
-          { defaultMessage: 'Google Workspace admin email' }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.gdrive.workspaceAdminEmail.label', {
+          defaultMessage: 'Google Workspace admin email',
+        }),
         options: [],
         order: 6,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
-          'searchConnectors.nativeConnectors.gdrive.workspaceAdminEmail.tooltip',
-          {
-            defaultMessage:
-              'In order to use Document Level Security you need to enable Google Workspace domain-wide delegation of authority for your service account. A service account with delegated authority can impersonate admin user with sufficient permissions to fetch all users and their corresponding permissions.',
-          }
-        ),
-        type: FieldType.STRING,
+        tooltip: translate('searchConnectors.nativeConnectors.gdrive.workspaceAdminEmail.tooltip', {
+          defaultMessage:
+            'In order to use Document Level Security you need to enable Google Workspace domain-wide delegation of authority for your service account. A service account with delegated authority can impersonate admin user with sufficient permissions to fetch all users and their corresponding permissions.',
+        }),
+        type: STRING,
         ui_restrictions: [],
         validations: [
           {
@@ -1683,19 +1656,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       max_concurrency: {
         default_value: 25,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.nativeConnectors.gdrive.maxHTTPRequest.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.gdrive.maxHTTPRequest.label', {
           defaultMessage: 'Maximum concurrent HTTP requests',
         }),
         options: [],
         order: 7,
         required: false,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.gdrive.maxHTTPRequest.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.gdrive.maxHTTPRequest.tooltip', {
           defaultMessage:
             'This setting determines the maximum number of concurrent HTTP requests sent to the Google API to fetch data. Increasing this value can improve data retrieval speed, but it may also place higher demands on system resources and network bandwidth.',
         }),
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [
           {
@@ -1708,28 +1681,28 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 8,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.google_drive.name', {
+    name: translate('searchConnectors.nativeConnectors.google_drive.name', {
       defaultMessage: 'Google Drive',
     }),
     serviceType: 'google_drive',
@@ -1739,25 +1712,25 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       data_source: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.dataSourceLabel', {
+        display: DROPDOWN,
+        label: translate('searchConnectors.nativeConnectors.jira.dataSourceLabel', {
           defaultMessage: 'Jira data source',
         }),
         options: [
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.jira.jiraCloudLabel', {
+            label: translate('searchConnectors.nativeConnectors.jira.jiraCloudLabel', {
               defaultMessage: 'Jira Cloud',
             }),
             value: 'jira_cloud',
           },
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.jira.jiraServerLabel', {
+            label: translate('searchConnectors.nativeConnectors.jira.jiraServerLabel', {
               defaultMessage: 'Jira Server',
             }),
             value: 'jira_server',
           },
           {
-            label: i18n.translate('searchConnectors.nativeConnectors.jira.jiraDataCenterLabel', {
+            label: translate('searchConnectors.nativeConnectors.jira.jiraDataCenterLabel', {
               defaultMessage: 'Jira Data Center',
             }),
             value: 'jira_data_center',
@@ -1767,7 +1740,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'jira_cloud',
@@ -1780,8 +1753,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'jira_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.serverUsername', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.jira.serverUsername', {
           defaultMessage: 'Jira Server username',
         }),
         options: [],
@@ -1789,7 +1762,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'admin',
@@ -1802,8 +1775,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'jira_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.serverPasswordLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.jira.serverPasswordLabel', {
           defaultMessage: 'Jira Server password',
         }),
         options: [],
@@ -1811,7 +1784,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'changeme',
@@ -1824,8 +1797,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'jira_data_center',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.dataCenterUsername', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.jira.dataCenterUsername', {
           defaultMessage: 'Username',
         }),
         options: [],
@@ -1833,7 +1806,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'admin',
@@ -1846,8 +1819,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'jira_data_center',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.dataCenterPasswordLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.jira.dataCenterPasswordLabel', {
           defaultMessage: 'Password',
         }),
         options: [],
@@ -1855,7 +1828,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'changeme',
@@ -1868,8 +1841,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'jira_cloud',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.cloudServiceAccountLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.jira.cloudServiceAccountLabel', {
           defaultMessage: 'Jira Cloud email address',
         }),
         options: [],
@@ -1877,14 +1850,11 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         placeholder: 'me@example.com',
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
-          'searchConnectors.nativeConnectors.jira.cloudServiceAccountTooltip',
-          {
-            defaultMessage:
-              'Email address associated with Jira Cloud account. E.g. jane.doe@gmail.com',
-          }
-        ),
-        type: FieldType.STRING,
+        tooltip: translate('searchConnectors.nativeConnectors.jira.cloudServiceAccountTooltip', {
+          defaultMessage:
+            'Email address associated with Jira Cloud account. E.g. jane.doe@gmail.com',
+        }),
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1897,8 +1867,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'jira_cloud',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.cloudApiTokenLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.jira.cloudApiTokenLabel', {
           defaultMessage: 'Jira Cloud API token',
         }),
         options: [],
@@ -1906,7 +1876,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'abc#123',
@@ -1914,8 +1884,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       jira_url: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.hostUrlLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.jira.hostUrlLabel', {
           defaultMessage: 'Jira host url',
         }),
         options: [],
@@ -1924,7 +1894,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1932,8 +1902,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       projects: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.jira.projectKeysLabel', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.jira.projectKeysLabel', {
           defaultMessage: 'Jira project keys',
         }),
         options: [],
@@ -1941,7 +1911,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: ADVANCED_RULES_IGNORED_LABEL,
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -1949,14 +1919,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_SSL_LABEL,
         options: [],
         order: 10,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -1969,14 +1939,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SSL_CERTIFICATE_LABEL,
         options: [],
         order: 11,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -1984,14 +1954,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 12,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -1999,14 +1969,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 100,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: MAX_CONCURRENT_DOWNLOADS_LABEL,
         options: [],
         order: 13,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [
           {
@@ -2019,18 +1989,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 14,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.jiraTooltip.name', {
+          translate('searchConnectors.nativeConnectors.jiraTooltip.name', {
             defaultMessage: 'Jira',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -2038,32 +2008,32 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 15,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.jira.name', {
+    name: translate('searchConnectors.nativeConnectors.jira.name', {
       defaultMessage: 'Jira',
     }),
     serviceType: 'jira',
@@ -2073,8 +2043,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tenant_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.microsoftTeams.tenantIdLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.microsoftTeams.tenantIdLabel', {
           defaultMessage: 'Tenant ID',
         }),
         options: [],
@@ -2082,7 +2052,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2090,8 +2060,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.microsoftTeams.clientIdLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.microsoftTeams.clientIdLabel', {
           defaultMessage: 'Client ID',
         }),
         options: [],
@@ -2099,7 +2069,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2107,8 +2077,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       secret_value: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.microsoftTeams.secretValueLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.microsoftTeams.secretValueLabel', {
           defaultMessage: 'Secret value',
         }),
         options: [],
@@ -2116,7 +2086,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2124,14 +2094,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       username: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: USERNAME_LABEL,
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2139,21 +2109,21 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: PASSWORD_LABEL,
         options: [],
         order: 5,
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
       },
     },
     features: {},
-    name: i18n.translate('searchConnectors.nativeConnectors.microsoftTeams.name', {
+    name: translate('searchConnectors.nativeConnectors.microsoftTeams.name', {
       defaultMessage: 'Microsoft Teams',
     }),
     serviceType: 'microsoft_teams',
@@ -2163,8 +2133,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       host: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.mongodb.configuration.hostLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mongodb.configuration.hostLabel', {
           defaultMessage: 'Server hostname',
         }),
         options: [],
@@ -2172,7 +2142,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2180,14 +2150,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       user: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: USERNAME_LABEL,
         options: [],
         order: 2,
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2195,14 +2165,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: PASSWORD_LABEL,
         options: [],
         order: 3,
         required: false,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2210,14 +2180,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       database: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: DATABASE_LABEL,
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2225,8 +2195,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       collection: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.mongodb.configuration.collectionLabel',
           {
             defaultMessage: 'Collection',
@@ -2237,7 +2207,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2245,8 +2215,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       direct_connection: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.mongodb.configuration.directConnectionLabel',
           {
             defaultMessage: 'Direct connection',
@@ -2257,7 +2227,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -2265,8 +2235,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.mongodb.configuration.sslEnabledLabel',
           {
             defaultMessage: 'SSL/TLS Connection',
@@ -2276,14 +2246,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 7,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.mongodb.configuration.sslEnabledTooltip',
           {
             defaultMessage:
               'This option establishes a secure connection to the MongoDB server using SSL/TLS encryption. Ensure that your MongoDB deployment supports SSL/TLS connections. Enable if MongoDB cluster uses DNS SRV records.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -2291,25 +2261,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_ca: {
         default_value: '',
         depends_on: [{ field: 'ssl_enabled', value: true }],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mongodb.configuration.sslCaLabel',
-          {
-            defaultMessage: 'Certificate Authority (.pem)',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mongodb.configuration.sslCaLabel', {
+          defaultMessage: 'Certificate Authority (.pem)',
+        }),
         options: [],
         order: 8,
         required: false,
         sensitive: false,
-        tooltip: i18n.translate(
-          'searchConnectors.nativeConnectors.mongodb.configuration.sslCaTooltip',
-          {
-            defaultMessage:
-              'Specifies the root certificate from the Certificate Authority. The value of the certificate is used to validate the certificate presented by the MongoDB instance.',
-          }
-        ),
-        type: FieldType.STRING,
+        tooltip: translate('searchConnectors.nativeConnectors.mongodb.configuration.sslCaTooltip', {
+          defaultMessage:
+            'Specifies the root certificate from the Certificate Authority. The value of the certificate is used to validate the certificate presented by the MongoDB instance.',
+        }),
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2317,8 +2281,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tls_insecure: {
         default_value: false,
         depends_on: [{ field: 'ssl_enabled', value: true }],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.mongodb.configuration.tlsInsecureLabel',
           {
             defaultMessage: 'Skip certificate verification',
@@ -2328,28 +2292,28 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 9,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.mongodb.configuration.tlsInsecureTooltip',
           {
             defaultMessage:
               "This option skips certificate validation for TLS/SSL connections to your MongoDB server. We strongly recommend setting this option to 'disable'.",
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.FILTERING_ADVANCED_CONFIG]: true,
-      [FeatureName.FILTERING_RULES]: true,
-      [FeatureName.SYNC_RULES]: {
+      [FILTERING_ADVANCED_CONFIG]: true,
+      [FILTERING_RULES]: true,
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.mongodb.name', {
+    name: translate('searchConnectors.nativeConnectors.mongodb.name', {
       defaultMessage: 'MongoDB',
     }),
     serviceType: 'mongodb',
@@ -2359,8 +2323,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       host: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.mssql.configuration.hostLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mssql.configuration.hostLabel', {
           defaultMessage: 'Host',
         }),
         options: [],
@@ -2368,7 +2332,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2376,14 +2340,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       port: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: PORT_LABEL,
         options: [],
         order: 2,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [],
         value: 1433,
@@ -2391,19 +2355,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       username: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mssql.configuration.usernameLabel',
-          {
-            defaultMessage: 'Username',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mssql.configuration.usernameLabel', {
+          defaultMessage: 'Username',
+        }),
         options: [],
         order: 3,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2411,19 +2372,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mssql.configuration.passwordLabel',
-          {
-            defaultMessage: 'Password',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mssql.configuration.passwordLabel', {
+          defaultMessage: 'Password',
+        }),
         options: [],
         order: 4,
         required: true,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2431,14 +2389,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       database: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: DATABASE_LABEL,
         options: [],
         order: 5,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2446,8 +2404,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tables: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.mssql.configuration.tablesLabel', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.mssql.configuration.tablesLabel', {
           defaultMessage: 'Comma-separated list of tables',
         }),
         options: [],
@@ -2455,7 +2413,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -2463,14 +2421,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_SSL_LABEL,
         options: [],
         order: 10,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -2478,14 +2436,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_ca: {
         default_value: '',
         depends_on: [{ field: 'ssl_enabled', value: true }],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SSL_CERTIFICATE_LABEL,
         options: [],
         order: 11,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2493,8 +2451,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       schema: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.mssql.configuration.schemaLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mssql.configuration.schemaLabel', {
           defaultMessage: 'Schema',
         }),
         options: [],
@@ -2502,7 +2460,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'dbo',
@@ -2510,19 +2468,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       fetch_size: {
         default_value: 50,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mssql.configuration.rowsFetchedLabel',
-          {
-            defaultMessage: 'Rows fetched per request',
-          }
-        ),
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.mssql.configuration.rowsFetchedLabel', {
+          defaultMessage: 'Rows fetched per request',
+        }),
         options: [],
         order: 7,
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 50,
@@ -2530,19 +2485,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mssql.configuration.retriesLabel',
-          {
-            defaultMessage: 'Retries per request',
-          }
-        ),
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.mssql.configuration.retriesLabel', {
+          defaultMessage: 'Retries per request',
+        }),
         options: [],
         order: 8,
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -2550,8 +2502,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       validate_host: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.mssql.configuration.validateHostLabel',
           {
             defaultMessage: 'Validate host',
@@ -2562,19 +2514,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.mssql.name', {
+    name: translate('searchConnectors.nativeConnectors.mssql.name', {
       defaultMessage: 'Microsoft SQL',
     }),
     serviceType: 'mssql',
@@ -2584,8 +2536,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       host: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.mysql.configuration.hostLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mysql.configuration.hostLabel', {
           defaultMessage: 'Host',
         }),
         options: [],
@@ -2593,7 +2545,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2601,14 +2553,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       port: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: PORT_LABEL,
         options: [],
         order: 2,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2616,19 +2568,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       user: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mysql.configuration.usernameLabel',
-          {
-            defaultMessage: 'Username',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mysql.configuration.usernameLabel', {
+          defaultMessage: 'Username',
+        }),
         options: [],
         order: 3,
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2636,19 +2585,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mysql.configuration.passwordLabel',
-          {
-            defaultMessage: 'Password',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.mysql.configuration.passwordLabel', {
+          defaultMessage: 'Password',
+        }),
         options: [],
         order: 4,
         required: false,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2656,14 +2602,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       database: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: DATABASE_LABEL,
         options: [],
         order: 5,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2671,8 +2617,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tables: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.nativeConnectors.mysql.configuration.tablesLabel', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.mysql.configuration.tablesLabel', {
           defaultMessage: 'Comma-separated list of tables',
         }),
         options: [],
@@ -2680,7 +2626,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -2688,14 +2634,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_SSL_LABEL,
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -2703,14 +2649,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_ca: {
         default_value: '',
         depends_on: [{ field: 'ssl_enabled', value: true }],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SSL_CERTIFICATE_LABEL,
         options: [],
         order: 8,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2718,19 +2664,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       fetch_size: {
         default_value: 50,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mysql.configuration.rowsFetchedLabel',
-          {
-            defaultMessage: 'Rows fetched per request',
-          }
-        ),
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.mysql.configuration.rowsFetchedLabel', {
+          defaultMessage: 'Rows fetched per request',
+        }),
         options: [],
         order: 9,
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 50,
@@ -2738,31 +2681,28 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.mysql.configuration.retriesLabel',
-          {
-            defaultMessage: 'Retries per request',
-          }
-        ),
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.mysql.configuration.retriesLabel', {
+          defaultMessage: 'Retries per request',
+        }),
         options: [],
         order: 10,
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.mysql.name', {
+    name: translate('searchConnectors.nativeConnectors.mysql.name', {
       defaultMessage: 'MySQL',
     }),
     serviceType: 'mysql',
@@ -2772,14 +2712,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       username: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: USERNAME_LABEL,
         options: [],
         order: 1,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2787,14 +2727,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: PASSWORD_LABEL,
         options: [],
         order: 2,
         required: true,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2802,8 +2742,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       server_ip: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.networkDrive.ipAddressLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.networkDrive.ipAddressLabel', {
           defaultMessage: 'IP address',
         }),
         options: [],
@@ -2812,7 +2752,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2820,14 +2760,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       server_port: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: PORT_LABEL,
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [],
         value: 445,
@@ -2835,8 +2775,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       drive_path: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.networkDrive.pathLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.networkDrive.pathLabel', {
           defaultMessage: 'Path',
         }),
         options: [],
@@ -2845,7 +2785,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2853,33 +2793,33 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 6,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.networkDriveTooltip.name', {
+          translate('searchConnectors.nativeConnectors.networkDriveTooltip.name', {
             defaultMessage: 'Network drive',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.networkDrive.name', {
+    name: translate('searchConnectors.nativeConnectors.networkDrive.name', {
       defaultMessage: 'Network drive',
     }),
     serviceType: 'network_drive',
@@ -2889,8 +2829,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       notion_secret_key: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.notion.notionSecretKeyLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.notion.notionSecretKeyLabel', {
           defaultMessage: 'Notion Secret Key',
         }),
         options: [],
@@ -2898,7 +2838,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2906,8 +2846,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       databases: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.notion.databasesLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.notion.databasesLabel', {
           defaultMessage: 'List of Databases',
         }),
         options: [],
@@ -2915,7 +2855,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2923,8 +2863,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       pages: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.notion.pagesLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.notion.pagesLabel', {
           defaultMessage: 'List of Pages',
         }),
         options: [],
@@ -2932,7 +2872,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -2940,19 +2880,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       index_comments: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate('searchConnectors.nativeConnectors.notion.indexCommentsLabel', {
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.notion.indexCommentsLabel', {
           defaultMessage: 'Enable indexing comments',
         }),
         options: [],
         order: 4,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.notion.indexCommentsTooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.notion.indexCommentsTooltip', {
           defaultMessage:
             'Enabling this will increase the amount of network calls to the source, and may decrease performance',
         }),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -2960,26 +2900,26 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 30,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: MAX_CONCURRENT_DOWNLOADS_LABEL,
         options: [],
         order: 5,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.notion.name', {
+    name: translate('searchConnectors.nativeConnectors.notion.name', {
       defaultMessage: 'Notion',
     }),
     serviceType: 'notion',
@@ -2989,8 +2929,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.oneDrive.azureClientId.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.oneDrive.azureClientId.label', {
           defaultMessage: 'Azure application Client ID',
         }),
         options: [],
@@ -2998,7 +2938,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3006,17 +2946,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_secret: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.oneDrive.azureClientSecret.label',
-          { defaultMessage: 'Azure application Client Secret' }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.oneDrive.azureClientSecret.label', {
+          defaultMessage: 'Azure application Client Secret',
+        }),
         options: [],
         order: 2,
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3024,8 +2963,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tenant_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.oneDrive.tenantId.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.oneDrive.tenantId.label', {
           defaultMessage: 'Azure application Tenant ID',
         }),
         options: [],
@@ -3033,7 +2972,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3041,14 +2980,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 4,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -3056,14 +2995,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 15,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: MAX_CONCURRENT_DOWNLOADS_LABEL,
         options: [],
         order: 5,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -3071,18 +3010,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 6,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.oneDriveTooltip.name', {
+          translate('searchConnectors.nativeConnectors.oneDriveTooltip.name', {
             defaultMessage: 'OneDrive',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -3090,31 +3029,31 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.onedrive.name', {
+    name: translate('searchConnectors.nativeConnectors.onedrive.name', {
       defaultMessage: 'OneDrive',
     }),
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
@@ -3125,8 +3064,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       host: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.oracle.configuration.hostLabel', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.oracle.configuration.hostLabel', {
           defaultMessage: 'Host',
         }),
         options: [],
@@ -3134,7 +3073,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3142,8 +3081,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       port: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.nativeConnectors.oracle.configuration.portLabel', {
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.oracle.configuration.portLabel', {
           defaultMessage: 'Port',
         }),
         options: [],
@@ -3151,7 +3090,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3159,19 +3098,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       username: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.oracle.configuration.usernameLabel',
-          {
-            defaultMessage: 'Username',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.oracle.configuration.usernameLabel', {
+          defaultMessage: 'Username',
+        }),
         options: [],
         order: 3,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3179,19 +3115,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.oracle.configuration.passwordLabel',
-          {
-            defaultMessage: 'Password',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.oracle.configuration.passwordLabel', {
+          defaultMessage: 'Password',
+        }),
         options: [],
         order: 4,
         required: true,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3199,19 +3132,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       database: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.oracle.configuration.databaseLabel',
-          {
-            defaultMessage: 'Database',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.oracle.configuration.databaseLabel', {
+          defaultMessage: 'Database',
+        }),
         options: [],
         order: 5,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3219,19 +3149,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tables: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.oracle.configuration.tablesLabel',
-          {
-            defaultMessage: 'Comma-separated list of tables',
-          }
-        ),
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.oracle.configuration.tablesLabel', {
+          defaultMessage: 'Comma-separated list of tables',
+        }),
         options: [],
         order: 6,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -3239,19 +3166,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       fetch_size: {
         default_value: 50,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.oracle.configuration.fetch_sizeLabel',
-          {
-            defaultMessage: 'Rows fetched per request',
-          }
-        ),
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.oracle.configuration.fetch_sizeLabel', {
+          defaultMessage: 'Rows fetched per request',
+        }),
         options: [],
         order: 7,
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 50,
@@ -3259,8 +3183,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
+        display: NUMERIC,
+        label: translate(
           'searchConnectors.nativeConnectors.oracle.configuration.retry_countLabel',
           {
             defaultMessage: 'Retries per request',
@@ -3271,7 +3195,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -3279,8 +3203,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       oracle_protocol: {
         default_value: 'TCP',
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate(
+        display: DROPDOWN,
+        label: translate(
           'searchConnectors.nativeConnectors.oracle.configuration.oracle_protocolLabel',
           {
             defaultMessage: 'Oracle connection protocol',
@@ -3300,7 +3224,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 'TCP',
@@ -3308,8 +3232,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       oracle_home: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.oracle.configuration.oracle_homeLabel',
           {
             defaultMessage: 'Path to Oracle Home',
@@ -3320,7 +3244,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -3328,8 +3252,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       wallet_configuration_path: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.oracle.configuration.wallet_configuration_pathLabel',
           {
             defaultMessage: 'Path to SSL Wallet configuration files',
@@ -3340,19 +3264,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: false },
         basic: { enabled: true },
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.oracle.name', {
+    name: translate('searchConnectors.nativeConnectors.oracle.name', {
       defaultMessage: 'Oracle',
     }),
     serviceType: 'oracle',
@@ -3362,8 +3286,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       data_source: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.data_source.label', {
+        display: DROPDOWN,
+        label: translate('searchConnectors.nativeConnectors.outlook.data_source.label', {
           defaultMessage: 'Outlook data source',
         }),
         options: [
@@ -3380,7 +3304,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'outlook_cloud',
@@ -3393,8 +3317,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_cloud',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.tenant_id.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.tenant_id.label', {
           defaultMessage: 'Tenant ID',
         }),
         options: [],
@@ -3402,7 +3326,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3415,8 +3339,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_cloud',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.client_id.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.client_id.label', {
           defaultMessage: 'Client ID',
         }),
         options: [],
@@ -3424,7 +3348,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3437,8 +3361,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_cloud',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.client_secret.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.client_secret.label', {
           defaultMessage: 'Client Secret Value',
         }),
         options: [],
@@ -3446,7 +3370,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3459,21 +3383,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.exchange_server.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.exchange_server.label', {
           defaultMessage: 'Exchange Server',
         }),
         options: [],
         order: 5,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
-          'searchConnectors.nativeConnectors.outlook.exchange_server.tooltip',
-          {
-            defaultMessage: "Exchange server's IP address. E.g. 127.0.0.1",
-          }
-        ),
-        type: FieldType.STRING,
+        tooltip: translate('searchConnectors.nativeConnectors.outlook.exchange_server.tooltip', {
+          defaultMessage: "Exchange server's IP address. E.g. 127.0.0.1",
+        }),
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3486,8 +3407,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.outlook.active_directory_server.label',
           {
             defaultMessage: 'Active Directory Server',
@@ -3497,13 +3418,13 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 6,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.outlook.active_directory_server.tooltip',
           {
             defaultMessage: "Active Directory server's IP address. E.g. 127.0.0.1",
           }
         ),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3516,8 +3437,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.username.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.username.label', {
           defaultMessage: 'Exchange server username',
         }),
         options: [],
@@ -3525,7 +3446,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3538,8 +3459,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.password.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.password.label', {
           defaultMessage: 'Exchange server password',
         }),
         options: [],
@@ -3547,7 +3468,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3560,18 +3481,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_server',
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.domain.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.domain.label', {
           defaultMessage: 'Exchange server domain name',
         }),
         options: [],
         order: 9,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.outlook.domain.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.outlook.domain.tooltip', {
           defaultMessage: 'Domain name such as gmail.com, outlook.com',
         }),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3584,8 +3505,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'outlook_server',
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.ssl_enabled.label', {
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.outlook.ssl_enabled.label', {
           defaultMessage: 'Enable SSL',
         }),
         options: [],
@@ -3593,7 +3514,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -3610,8 +3531,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.outlook.ssl_ca.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.outlook.ssl_ca.label', {
           defaultMessage: 'SSL certificate',
         }),
         options: [],
@@ -3619,7 +3540,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3627,8 +3548,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.outlook.use_text_extraction_service.label',
           {
             defaultMessage: 'Use text extraction service',
@@ -3638,14 +3559,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 12,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.outlook.use_text_extraction_service.toolip',
           {
             defaultMessage:
               'Requires a separate deployment of the Elastic Text Extraction Service. Requires that pipeline settings disable text extraction.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
@@ -3653,36 +3574,36 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 13,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.outlookTooltip.name', {
+          translate('searchConnectors.nativeConnectors.outlookTooltip.name', {
             defaultMessage: 'Outlook',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: false },
         basic: { enabled: true },
       },
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.outlook.name', {
+    name: translate('searchConnectors.nativeConnectors.outlook.name', {
       defaultMessage: 'Outlook',
     }),
     serviceType: 'outlook',
@@ -3692,19 +3613,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       host: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.postgresql.configuration.hostLabel',
-          {
-            defaultMessage: 'Host',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.postgresql.configuration.hostLabel', {
+          defaultMessage: 'Host',
+        }),
         options: [],
         order: 1,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3712,14 +3630,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       port: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: PORT_LABEL,
         options: [],
         order: 2,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [],
         value: 5432,
@@ -3727,14 +3645,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       username: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: USERNAME_LABEL,
         options: [],
         order: 3,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3742,14 +3660,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: PASSWORD_LABEL,
         options: [],
         order: 4,
         required: true,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3757,14 +3675,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       database: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: DATABASE_LABEL,
         options: [],
         order: 5,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3772,14 +3690,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       schema: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SCHEMA_LABEL,
         options: [],
         order: 6,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3787,19 +3705,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tables: {
         default_value: '',
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.postgresql.configuration.tablesLabel',
-          {
-            defaultMessage: 'Comma-separated list of tables',
-          }
-        ),
+        display: TEXTAREA,
+        label: translate('searchConnectors.nativeConnectors.postgresql.configuration.tablesLabel', {
+          defaultMessage: 'Comma-separated list of tables',
+        }),
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -3807,14 +3722,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_SSL_LABEL,
         options: [],
         order: 10,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -3822,14 +3737,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_ca: {
         default_value: '',
         depends_on: [{ field: 'ssl_enabled', value: true }],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SSL_CERTIFICATE_LABEL,
         options: [],
         order: 11,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3837,8 +3752,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       fetch_size: {
         default_value: 50,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
+        display: NUMERIC,
+        label: translate(
           'searchConnectors.nativeConnectors.postgresql.configuration.rowsFetchedLabel',
           {
             defaultMessage: 'Rows fetched per request',
@@ -3849,7 +3764,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 50,
@@ -3857,8 +3772,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate(
+        display: NUMERIC,
+        label: translate(
           'searchConnectors.nativeConnectors.postgresql.configuration.retriesLabel',
           {
             defaultMessage: 'Retries per request',
@@ -3869,19 +3784,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: '',
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.postgresql.name', {
+    name: translate('searchConnectors.nativeConnectors.postgresql.name', {
       defaultMessage: 'PostgreSQL',
     }),
     serviceType: 'postgresql',
@@ -3891,18 +3806,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       buckets: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate('searchConnectors.content.nativeConnectors.s3.buckets.label', {
+        display: TEXTAREA,
+        label: translate('searchConnectors.content.nativeConnectors.s3.buckets.label', {
           defaultMessage: 'AWS Buckets',
         }),
         options: [],
         order: 1,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.content.nativeConnectors.s3.buckets.tooltip', {
+        tooltip: translate('searchConnectors.content.nativeConnectors.s3.buckets.tooltip', {
           defaultMessage: 'AWS Buckets are ignored when Advanced Sync Rules are used.',
         }),
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3910,8 +3825,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       aws_access_key_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.content.nativeConnectors.s3.accessKey.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.content.nativeConnectors.s3.accessKey.label', {
           defaultMessage: 'AWS Access Key Id',
         }),
         options: [],
@@ -3919,7 +3834,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3927,8 +3842,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       aws_secret_access_key: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.content.nativeConnectors.s3.secretKey.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.content.nativeConnectors.s3.secretKey.label', {
           defaultMessage: 'AWS Secret Key',
         }),
         options: [],
@@ -3936,7 +3851,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -3944,9 +3859,9 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       read_timeout: {
         default_value: 90,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
 
-        label: i18n.translate('searchConnectors.content.nativeConnectors.s3.readTimeout.label', {
+        label: translate('searchConnectors.content.nativeConnectors.s3.readTimeout.label', {
           defaultMessage: 'Read timeout',
         }),
         options: [],
@@ -3954,7 +3869,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -3962,8 +3877,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       connect_timeout: {
         default_value: 90,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.content.nativeConnectors.s3.connectTimeout.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.content.nativeConnectors.s3.connectTimeout.label', {
           defaultMessage: 'Connection timeout',
         }),
         options: [],
@@ -3971,7 +3886,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -3979,8 +3894,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       max_attempts: {
         default_value: 5,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.content.nativeConnectors.s3.maxAttempts.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.content.nativeConnectors.s3.maxAttempts.label', {
           defaultMessage: 'Maximum retry attempts',
         }),
         options: [],
@@ -3988,7 +3903,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -3996,8 +3911,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       page_size: {
         default_value: 100,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.content.nativeConnectors.s3.maxPageSize.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.content.nativeConnectors.s3.maxPageSize.label', {
           defaultMessage: 'Maximum size of page',
         }),
         options: [],
@@ -4005,7 +3920,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: '',
@@ -4013,37 +3928,37 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 8,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
     },
-    name: i18n.translate('searchConnectors.content.nativeConnectors.s3.name', {
+    name: translate('searchConnectors.content.nativeConnectors.s3.name', {
       defaultMessage: 'S3',
     }),
     serviceType: 's3',
   },
   salesforce: {
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
@@ -4051,22 +3966,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       domain: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.content.nativeConnectors.salesforce.domain.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.content.nativeConnectors.salesforce.domain.label', {
           defaultMessage: 'Domain',
         }),
         options: [],
         order: 1,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
-          'searchConnectors.content.nativeConnectors.salesforce.domain.tooltip',
-          {
-            defaultMessage:
-              "The domain for your Salesforce instance. If your Salesforce URL is 'https://foo.salesforce.com', the domain would be 'foo'.",
-          }
-        ),
-        type: FieldType.STRING,
+        tooltip: translate('searchConnectors.content.nativeConnectors.salesforce.domain.tooltip', {
+          defaultMessage:
+            "The domain for your Salesforce instance. If your Salesforce URL is 'https://foo.salesforce.com', the domain would be 'foo'.",
+        }),
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4074,25 +3986,22 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.content.nativeConnectors.salesforce.clientId.label',
-          {
-            defaultMessage: 'Client ID',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.content.nativeConnectors.salesforce.clientId.label', {
+          defaultMessage: 'Client ID',
+        }),
         options: [],
         order: 2,
         required: true,
         sensitive: true,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.content.nativeConnectors.salesforce.clientId.tooltip',
           {
             defaultMessage:
               "The client id for your OAuth2-enabled connected app. Also called 'consumer key'",
           }
         ),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4100,8 +4009,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_secret: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.content.nativeConnectors.salesforce.clientSecret.label',
           { defaultMessage: 'Client Secret' }
         ),
@@ -4109,14 +4018,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 3,
         required: true,
         sensitive: true,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.content.nativeConnectors.salesforce.clientSecret.tooltip',
           {
             defaultMessage:
               "The client secret for your OAuth2-enabled connected app. Also called 'consumer secret'",
           }
         ),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4124,18 +4033,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.salesforce.name', {
+          translate('searchConnectors.nativeConnectors.salesforce.name', {
             defaultMessage: 'Salesforce',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -4143,20 +4052,20 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 5,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
-    name: i18n.translate('searchConnectors.content.nativeConnectors.salesforce.name', {
+    name: translate('searchConnectors.content.nativeConnectors.salesforce.name', {
       defaultMessage: 'Salesforce',
     }),
     serviceType: 'salesforce',
@@ -4166,8 +4075,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       url: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.servicenow.url.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.servicenow.url.label', {
           defaultMessage: 'Service URL',
         }),
         options: [],
@@ -4175,7 +4084,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'http://127.0.0.1:9318',
@@ -4183,8 +4092,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       username: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.servicenow.username.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.servicenow.username.label', {
           defaultMessage: 'Username',
         }),
         options: [],
@@ -4192,7 +4101,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4200,8 +4109,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.servicenow.password.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.servicenow.password.label', {
           defaultMessage: 'Password',
         }),
         options: [],
@@ -4209,7 +4118,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4217,18 +4126,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       services: {
         default_value: null,
         depends_on: [],
-        label: i18n.translate('searchConnectors.nativeConnectors.servicenow.services.label', {
+        label: translate('searchConnectors.nativeConnectors.servicenow.services.label', {
           defaultMessage: 'Comma-separated list of services',
         }),
-        display: DisplayType.TEXTAREA,
+        display: TEXTAREA,
         options: [],
         order: 4,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.servicenow.services.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.servicenow.services.tooltip', {
           defaultMessage: 'List of services is ignored when Advanced Sync Rules are used.',
         }),
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -4236,14 +4145,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 5,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -4251,14 +4160,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       concurrent_downloads: {
         default_value: 10,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: MAX_CONCURRENT_DOWNLOADS_LABEL,
         options: [],
         order: 6,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 10,
@@ -4266,14 +4175,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
@@ -4281,35 +4190,35 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 8,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.servicenow.configuration.useDocumentLevelSecurityTooltip',
           {
             defaultMessage:
               'Document level security ensures identities and permissions set in ServiceNow are maintained in Elasticsearch. This enables you to restrict and personalize read-access users and groups have to documents in this index. Access control syncs ensure this metadata is kept up to date in your Elasticsearch documents.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.servicenow.name', {
+    name: translate('searchConnectors.nativeConnectors.servicenow.name', {
       defaultMessage: 'ServiceNow',
     }),
     serviceType: 'servicenow',
@@ -4319,8 +4228,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tenant_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.tenantIdLabel',
           {
             defaultMessage: 'Tenant ID',
@@ -4331,7 +4240,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4339,8 +4248,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       tenant_name: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.tenantNameLabel',
           {
             defaultMessage: 'Tenant name',
@@ -4351,7 +4260,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4359,8 +4268,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.clientIdLabel',
           {
             defaultMessage: 'Client ID',
@@ -4371,7 +4280,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4379,8 +4288,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       secret_value: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.secretValueLabel',
           {
             defaultMessage: 'Secret value',
@@ -4391,7 +4300,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4399,8 +4308,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       site_collections: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate(
+        display: TEXTAREA,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.siteCollectionsLabel',
           {
             defaultMessage: 'Comma-separated list of sites',
@@ -4410,7 +4319,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 5,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.siteCollectionsTooltip',
           {
             defaultMessage:
@@ -4418,7 +4327,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
               'Use * to include all available sites.',
           }
         ),
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '*',
@@ -4426,14 +4335,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 6,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
@@ -4441,8 +4350,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: false,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.useDocumentLevelSecurityLabel',
           {
             defaultMessage: 'Enable document level security',
@@ -4452,14 +4361,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 7,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.useDocumentLevelSecurityTooltip',
           {
             defaultMessage:
               'Document level security ensures identities and permissions set in Sharepoint Online are maintained in Elasticsearch. This metadata is added to your Elasticsearch documents, so you can control user and group read-access. Access control syncs ensure this metadata is kept up to date.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -4472,8 +4381,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchDriveItemPermissionsLabel',
           {
             defaultMessage: 'Fetch drive item permissions',
@@ -4483,14 +4392,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 8,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchDriveItemPermissionsTooltip',
           {
             defaultMessage:
               'Enable this option to fetch drive item specific permissions. This setting can increase sync time.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
@@ -4503,8 +4412,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchUniquePagePermissionsLabel',
           {
             defaultMessage: 'Fetch unique page permissions',
@@ -4514,14 +4423,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 9,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchUniquePagePermissionsTooltip',
           {
             defaultMessage:
               'Enable this option to fetch unique page permissions. This setting can increase sync time. If this setting is disabled a page will inherit permissions from its parent site.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
@@ -4534,8 +4443,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchUniqueListPermissionsLabel',
           {
             defaultMessage: 'Fetch unique list permissions',
@@ -4545,14 +4454,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 10,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchUniqueListPermissionsTooltip',
           {
             defaultMessage:
               'Enable this option to fetch unique list permissions. This setting can increase sync time. If this setting is disabled a list will inherit permissions from its parent site.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
@@ -4565,8 +4474,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchUniqueListItemPermissionsLabel',
           {
             defaultMessage: 'Fetch unique list item permissions',
@@ -4576,14 +4485,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 11,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchUniqueListItemPermissionsTooltip',
           {
             defaultMessage:
               'Enable this option to fetch unique list item permissions. This setting can increase sync time. If this setting is disabled a list item will inherit permissions from its parent site.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
@@ -4591,8 +4500,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       enumerate_all_sites: {
         default_value: true,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.enumerateAllSitesLabel',
           { defaultMessage: 'Enumerate all sites?' }
         ),
@@ -4600,14 +4509,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 6,
         required: false,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.enumerateAllSitesTooltip',
           {
             defaultMessage:
               'If enabled, sites will be fetched in bulk, then filtered down to the configured list of sites. This is efficient when syncing many sites. If disabled, each configured site will be fetched with an individual request. This is efficient when syncing fewer sites.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
@@ -4615,8 +4524,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       fetch_subsites: {
         default_value: false,
         depends_on: [{ field: 'enumerate_all_sites', value: false }],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchSubsitesLabel',
           {
             defaultMessage: 'Fetch sub-sites of configured sites?',
@@ -4626,32 +4535,32 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 7,
         required: false,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_online.configuration.fetchSubsitesTooltip',
           {
             defaultMessage:
               'Whether subsites of the configured site(s) should be automatically fetched.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: true },
         basic: { enabled: true },
       },
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.sharepoint_online.name', {
+    name: translate('searchConnectors.nativeConnectors.sharepoint_online.name', {
       defaultMessage: 'Sharepoint Online',
     }),
     serviceType: 'sharepoint_online',
@@ -4661,8 +4570,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       authentication: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.DROPDOWN,
-        label: i18n.translate(
+        display: DROPDOWN,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.authentication',
           {
             defaultMessage: 'Authentication',
@@ -4670,7 +4579,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         ),
         options: [
           {
-            label: i18n.translate(
+            label: translate(
               'searchConnectors.nativeConnectors.sharepoint_server.options.basicLabel',
               {
                 defaultMessage: 'Basic',
@@ -4679,7 +4588,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: 'basic_auth',
           },
           {
-            label: i18n.translate(
+            label: translate(
               'searchConnectors.nativeConnectors.sharepoint_server.options.ntlmLabel',
               {
                 defaultMessage: 'NTLM',
@@ -4692,7 +4601,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: 'basic_auth',
@@ -4700,8 +4609,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       username: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.username',
           {
             defaultMessage: 'SharePoint Server username',
@@ -4712,7 +4621,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4720,8 +4629,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       password: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
+        display: TEXTBOX,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.password',
           {
             defaultMessage: 'SharePoint Server password',
@@ -4732,7 +4641,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4740,19 +4649,16 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       host_url: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.sharepoint_server.configuration.host',
-          {
-            defaultMessage: 'SharePoint host',
-          }
-        ),
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.sharepoint_server.configuration.host', {
+          defaultMessage: 'SharePoint host',
+        }),
         options: [],
         order: 4,
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4760,8 +4666,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       site_collections: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTAREA,
-        label: i18n.translate(
+        display: TEXTAREA,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.site_collections',
           {
             defaultMessage: 'Comma-separated list of SharePoint site collections to index',
@@ -4772,7 +4678,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: '',
-        type: FieldType.LIST,
+        type: LIST,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4780,14 +4686,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       ssl_enabled: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_SSL_LABEL,
         options: [],
         order: 6,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -4800,14 +4706,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TEXTBOX,
+        display: TEXTBOX,
         label: SSL_CERTIFICATE_LABEL,
         options: [],
         order: 7,
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4815,14 +4721,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       retry_count: {
         default_value: 3,
         depends_on: [],
-        display: DisplayType.NUMERIC,
+        display: NUMERIC,
         label: RETRIES_PER_REQUEST_LABEL,
         options: [],
         order: 8,
         required: false,
         sensitive: false,
         tooltip: null,
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: ['advanced'],
         validations: [],
         value: 3,
@@ -4830,14 +4736,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 9,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
@@ -4845,18 +4751,18 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_document_level_security: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: ENABLE_DOCUMENT_LEVEL_SECURITY_LABEL,
         options: [],
         order: 10,
         required: true,
         sensitive: false,
         tooltip: getEnableDocumentLevelSecurityTooltip(
-          i18n.translate('searchConnectors.nativeConnectors.sharepoint_server.name', {
+          translate('searchConnectors.nativeConnectors.sharepoint_server.name', {
             defaultMessage: 'Sharepoint Server',
           })
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -4869,8 +4775,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.fetchUniqueListPermissionsLabel',
           {
             defaultMessage: 'Fetch unique list permissions',
@@ -4880,14 +4786,14 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 11,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.fetchUniqueListPermissionsTooltip',
           {
             defaultMessage:
               'Enable this option to fetch unique list permissions. This setting can increase sync time. If this setting is disabled a list will inherit permissions from its parent site.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
@@ -4900,8 +4806,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
             value: true,
           },
         ],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
+        display: TOGGLE,
+        label: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.fetchUniqueListItemPermissionsLabel',
           {
             defaultMessage: 'Fetch unique list item permissions',
@@ -4911,32 +4817,32 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         order: 12,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.sharepoint_server.configuration.fetchUniqueListItemPermissionsTooltip',
           {
             defaultMessage:
               'Enable this option to fetch unique list item permissions. This setting can increase sync time. If this setting is disabled a list item will inherit permissions from its parent site.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
       },
     },
     features: {
-      [FeatureName.SYNC_RULES]: {
+      [SYNC_RULES]: {
         advanced: { enabled: false },
         basic: { enabled: true },
       },
-      [FeatureName.DOCUMENT_LEVEL_SECURITY]: {
+      [DOCUMENT_LEVEL_SECURITY]: {
         enabled: true,
       },
-      [FeatureName.INCREMENTAL_SYNC]: {
+      [INCREMENTAL_SYNC]: {
         enabled: true,
       },
     },
-    name: i18n.translate('searchConnectors.nativeConnectors.sharepoint_server.name', {
+    name: translate('searchConnectors.nativeConnectors.sharepoint_server.name', {
       defaultMessage: 'Sharepoint Server',
     }),
     serviceType: 'sharepoint_server',
@@ -4946,19 +4852,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       token: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.slack.token.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.slack.token.label', {
           defaultMessage: 'Authentication Token',
         }),
         options: [],
         order: 1,
         required: true,
         sensitive: true,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.slack.token.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.slack.token.tooltip', {
           defaultMessage:
             'The Slack Authentication Token for the slack application you created. See the docs for details.',
         }),
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4966,19 +4872,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       fetch_last_n_days: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.nativeConnectors.slack.fetchLastNDays.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.slack.fetchLastNDays.label', {
           defaultMessage: 'Days of message history to fetch',
         }),
         options: [],
         order: 2,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.slack.fetchLastNDays.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.slack.fetchLastNDays.tooltip', {
           defaultMessage:
             'How far back in time to request message history from slack. Messages older than this will not be indexed.',
         }),
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -4986,22 +4892,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       auto_join_channels: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate('searchConnectors.nativeConnectors.slack.autoJoinChannels.label', {
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.slack.autoJoinChannels.label', {
           defaultMessage: 'Automatically join channels',
         }),
         options: [],
         order: 3,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
-          'searchConnectors.nativeConnectors.slack.autoJoinChannels.tooltip',
-          {
-            defaultMessage:
-              'The Slack application bot will only be able to read conversation history from channels it has joined. The default requires it to be manually invited to channels. Enabling this allows it to automatically invite itself into all public channels.',
-          }
-        ),
-        type: FieldType.BOOLEAN,
+        tooltip: translate('searchConnectors.nativeConnectors.slack.autoJoinChannels.tooltip', {
+          defaultMessage:
+            'The Slack application bot will only be able to read conversation history from channels it has joined. The default requires it to be manually invited to channels. Enabling this allows it to automatically invite itself into all public channels.',
+        }),
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -5009,26 +4912,26 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       sync_users: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate('searchConnectors.nativeConnectors.slack.syncUsers.label', {
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.slack.syncUsers.label', {
           defaultMessage: 'Sync users',
         }),
         options: [],
         order: 4,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.slack.syncUsers.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.slack.syncUsers.tooltip', {
           defaultMessage:
             'Whether or not Slack Users should be indexed as documents in Elasticsearch.',
         }),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: true,
       },
     },
     features: {},
-    name: i18n.translate('searchConnectors.nativeConnectors.slack.name', {
+    name: translate('searchConnectors.nativeConnectors.slack.name', {
       defaultMessage: 'Slack',
     }),
     serviceType: 'slack',
@@ -5038,8 +4941,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       account_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.zoom.accountId.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.zoom.accountId.label', {
           defaultMessage: 'Account ID',
         }),
         options: [],
@@ -5047,7 +4950,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -5055,8 +4958,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_id: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.zoom.clientId.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.zoom.clientId.label', {
           defaultMessage: 'Client ID',
         }),
         options: [],
@@ -5064,7 +4967,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: false,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -5072,8 +4975,8 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       client_secret: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TEXTBOX,
-        label: i18n.translate('searchConnectors.nativeConnectors.zoom.clientSecret.label', {
+        display: TEXTBOX,
+        label: translate('searchConnectors.nativeConnectors.zoom.clientSecret.label', {
           defaultMessage: 'Client secret',
         }),
         options: [],
@@ -5081,7 +4984,7 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
         required: true,
         sensitive: true,
         tooltip: null,
-        type: FieldType.STRING,
+        type: STRING,
         ui_restrictions: [],
         validations: [],
         value: '',
@@ -5089,25 +4992,22 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       fetch_past_meeting_details: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
-        label: i18n.translate(
-          'searchConnectors.nativeConnectors.zoom.fetchPastMeetingDetails.label',
-          {
-            defaultMessage: 'Fetch past meeting details',
-          }
-        ),
+        display: TOGGLE,
+        label: translate('searchConnectors.nativeConnectors.zoom.fetchPastMeetingDetails.label', {
+          defaultMessage: 'Fetch past meeting details',
+        }),
         options: [],
         order: 4,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate(
+        tooltip: translate(
           'searchConnectors.nativeConnectors.zoom.fetchPastMeetingDetails.tooltip',
           {
             defaultMessage:
               'Enable this option to fetch past past meeting details. This setting can increase sync time.',
           }
         ),
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: [],
         validations: [],
         value: false,
@@ -5115,19 +5015,19 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       recording_age: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.NUMERIC,
-        label: i18n.translate('searchConnectors.nativeConnectors.zoom.recordingAge.label', {
+        display: NUMERIC,
+        label: translate('searchConnectors.nativeConnectors.zoom.recordingAge.label', {
           defaultMessage: 'Recording Age Limit (Months)',
         }),
         options: [],
         order: 5,
         required: true,
         sensitive: false,
-        tooltip: i18n.translate('searchConnectors.nativeConnectors.zoom.recordingAge.tooltip', {
+        tooltip: translate('searchConnectors.nativeConnectors.zoom.recordingAge.tooltip', {
           defaultMessage:
             'How far back in time to request recordings from zoom. Recordings older than this will not be indexed.',
         }),
-        type: FieldType.INTEGER,
+        type: INTEGER,
         ui_restrictions: [],
         validations: [
           {
@@ -5140,21 +5040,21 @@ export const NATIVE_CONNECTOR_DEFINITIONS: Record<string, NativeConnector | unde
       use_text_extraction_service: {
         default_value: null,
         depends_on: [],
-        display: DisplayType.TOGGLE,
+        display: TOGGLE,
         label: USE_TEXT_EXTRACTION_SERVICE_LABEL,
         options: [],
         order: 6,
         required: true,
         sensitive: false,
         tooltip: USE_TEXT_EXTRACTION_SERVICE_TOOLTIP,
-        type: FieldType.BOOLEAN,
+        type: BOOLEAN,
         ui_restrictions: ['advanced'],
         validations: [],
         value: false,
       },
     },
     features: {},
-    name: i18n.translate('searchConnectors.nativeConnectors.zoom.name', {
+    name: translate('searchConnectors.nativeConnectors.zoom.name', {
       defaultMessage: 'Zoom',
     }),
     serviceType: 'zoom',
