+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/new_index/method_connector/new_connector_logic.ts
@@ -7,8 +7,7 @@
 
 import { kea, MakeLogicType } from 'kea';
 
-import { Connector } from '@kbn/search-connectors';
-import { ConnectorDefinition } from '@kbn/search-connectors-plugin/public';
+import { Connector, ConnectorDefinition } from '@kbn/search-connectors';
 
 import { Status } from '../../../../../../common/types/api';
 import { Actions } from '../../../../shared/api_logic/create_api_logic';
