+++ x-pack/plugins/enterprise_search/server/utils/search_result_provider.ts
@@ -10,7 +10,7 @@ import { takeUntil, of, map } from 'rxjs';
 import { GlobalSearchResultProvider } from '@kbn/global-search-plugin/server';
 import { i18n } from '@kbn/i18n';
 
-import { ConnectorServerSideDefinition } from '@kbn/search-connectors-plugin/server';
+import { ConnectorServerSideDefinition } from '@kbn/search-connectors';
 
 import { ConfigType } from '..';
 import {
