+++ x-pack/plugins/search_connectors/public/types.ts
@@ -5,8 +5,7 @@
  * 2.0.
  */
 
-import { ConnectorServerSideDefinition } from '../common/connectors';
-import { ConnectorDefinition } from '../common/types';
+import { ConnectorDefinition, ConnectorServerSideDefinition } from '@kbn/search-connectors';
 
 /* eslint-disable @typescript-eslint/no-empty-interface */
