+++ x-pack/plugins/search_connectors/server/types.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { ConnectorServerSideDefinition } from '../common/connectors';
+import { ConnectorServerSideDefinition } from '@kbn/search-connectors';
 
 /* eslint-disable @typescript-eslint/no-empty-interface */
