+++ x-pack/plugins/search_connectors/tsconfig.json
@@ -17,8 +17,7 @@
   "kbn_references": [
     "@kbn/core",
     "@kbn/config-schema",
-    "@kbn/doc-links",
     "@kbn/core-http-browser",
-    "@kbn/i18n",
+    "@kbn/search-connectors",
   ]
 }
